ReccomendationEngine/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
numpy==1.24.3
pandas==1.5.3
python_Levenshtein==0.23.0
scikit_learn==1.3.0
